+++ spec/pie-chart-spec.js
@@ -157,10 +157,11 @@ describe('dc.pieChart', function () {
             });
         });
         it('slice path fill should be set correctly', function () {
-            expect(d3.select(chart.selectAll('g.pie-slice path').nodes()[0]).attr('fill')).toMatch(/#3182bd/i);
-            expect(d3.select(chart.selectAll('g.pie-slice path').nodes()[1]).attr('fill')).toMatch(/#6baed6/i);
-            expect(d3.select(chart.selectAll('g.pie-slice path').nodes()[2]).attr('fill')).toMatch(/#9ecae1/i);
-            expect(d3.select(chart.selectAll('g.pie-slice path').nodes()[3]).attr('fill')).toMatch(/#c6dbef/i);
+            var numSlices = 5;
+            for (var i = 0; i < numSlices; i++) {
+                expect(d3.select(chart.selectAll('g.pie-slice path').nodes()[i]).attr('fill'))
+                    .toMatchColor(dc.config.defaultColors()[i]);
+            }
         });
         it('slice label should be created', function () {
             expect(chart.selectAll('svg text.pie-slice').data().length).toEqual(5);