+++ spec/row-chart-spec.js
@@ -117,11 +117,10 @@ describe('dc.rowChart', function () {
                 });
 
                 it('should fill each row rect with pre-defined colors', function () {
-                    expect(d3.select(chart.selectAll('g.row rect').nodes()[0]).attr('fill')).toMatch(/#3182bd/i);
-                    expect(d3.select(chart.selectAll('g.row rect').nodes()[1]).attr('fill')).toMatch(/#6baed6/i);
-                    expect(d3.select(chart.selectAll('g.row rect').nodes()[2]).attr('fill')).toMatch(/#9ecae1/i);
-                    expect(d3.select(chart.selectAll('g.row rect').nodes()[3]).attr('fill')).toMatch(/#c6dbef/i);
-                    expect(d3.select(chart.selectAll('g.row rect').nodes()[4]).attr('fill')).toMatch(/#e6550d/i);
+                    for (var i = 0; i < N; i++) {
+                        expect(d3.select(chart.selectAll('g.row rect').nodes()[i]).attr('fill'))
+                            .toMatchColor(dc.config.defaultColors()[i]);
+                    }
                 });
 
                 it('should create a row label from the data for each row', function () {